+++ packages/gatsby-image/src/index.js
@@ -79,21 +79,26 @@ const noscriptImg = props => {
   // HTML validation issues caused by empty values like width="" and height=""
   const src = props.src ? `src="${props.src}" ` : `src="" ` // required attribute
   const sizes = props.sizes ? `sizes="${props.sizes}" ` : ``
-  const srcSetWebp = props.srcSetWebp ? `<source type='image/webp' srcSet="${props.srcSetWebp}" ${sizes}/>` : ``
-  const srcSet = props.srcSet ? `<source srcSet="${props.srcSet}" ${sizes}/>` : ``
+  const srcSetWebp = props.srcSetWebp
+    ? `<source type='image/webp' srcSet="${props.srcSetWebp}" ${sizes}/>`
+    : ``
+  const srcSet = props.srcSet
+    ? `<source srcSet="${props.srcSet}" ${sizes}/>`
+    : ``
   const title = props.title ? `title="${props.title}" ` : ``
   const alt = props.alt ? `alt="${props.alt}" ` : `alt="" ` // required attribute
   const width = props.width ? `width="${props.width}" ` : ``
   const height = props.height ? `height="${props.height}" ` : ``
   const opacity = props.opacity ? props.opacity : `1`
   const transitionDelay = props.transitionDelay ? props.transitionDelay : `0.5s`
-  return (`<picture>${srcSetWebp}${srcSet}<img ${width}${height}${src}${alt}${title}style="position:absolute;top:0;left:0;transition:opacity 0.5s;transition-delay:${transitionDelay};opacity:${opacity};width:100%;height:100%;object-fit:cover;object-position:center"/></picture>`)
+  return `<picture>${srcSetWebp}${srcSet}<img ${width}${height}${src}${alt}${title}style="position:absolute;top:0;left:0;transition:opacity 0.5s;transition-delay:${transitionDelay};opacity:${opacity};width:100%;height:100%;object-fit:cover;object-position:center"/></picture>`
 }
 
 const Img = React.forwardRef((props, ref) => {
   const { style, onLoad, onError, ...otherProps } = props
 
-  return <img
+  return (
+    <img
       {...otherProps}
       onLoad={onLoad}
       onError={onError}
@@ -109,6 +114,7 @@ const Img = React.forwardRef((props, ref) => {
         ...style,
       }}
     />
+  )
 })
 
 Img.propTypes = {
@@ -154,7 +160,7 @@ class Image extends React.Component {
       IOSupported = false
     }
 
-    const hasNoScript =  !(this.props.critical && !this.props.fadeIn)
+    const hasNoScript = !(this.props.critical && !this.props.fadeIn)
 
     this.state = {
       isVisible,
@@ -199,7 +205,7 @@ class Image extends React.Component {
 
   handleImageLoaded() {
     activateCacheForImage(this.props)
-    
+
     this.setState({ imgLoaded: true })
     if (this.state.seenBefore) {
       this.setState({ fadeIn: false })
@@ -215,18 +221,15 @@ class Image extends React.Component {
       style = {},
       imgStyle = {},
       placeholderStyle = {},
+      placeholderClassName,
       fluid,
       fixed,
       backgroundColor,
       Tag,
     } = convertProps(this.props)
 
-    let bgColor
-    if (typeof backgroundColor === `boolean`) {
-      bgColor = `lightgray`
-    } else {
-      bgColor = backgroundColor
-    }
+    const bgColor =
+      typeof backgroundColor === `boolean` ? `lightgray` : backgroundColor
 
     const imagePlaceholderStyle = {
       opacity: this.state.imgLoaded ? 0 : 1,
@@ -242,6 +245,13 @@ class Image extends React.Component {
       ...imgStyle,
     }
 
+    const placeholderImageProps = {
+      title,
+      alt: !this.state.isVisible ? alt : ``,
+      style: imagePlaceholderStyle,
+      className: placeholderClassName,
+    }
+
     if (fluid) {
       const image = fluid
 
@@ -254,6 +264,7 @@ class Image extends React.Component {
             ...style,
           }}
           ref={this.handleRef}
+          key={`fluid-${JSON.stringify(image.srcSet)}`}
         >
           {/* Preserve the aspect ratio. */}
           <Tag
@@ -263,77 +274,66 @@ class Image extends React.Component {
             }}
           />
 
-            {/* Show the blurry base64 image. */}
-            {image.base64 && (
-              <Img
-                alt={alt}
-                title={title}
-                src={image.base64}
-                style={imagePlaceholderStyle}
-              />
-            )}
+          {/* Show the blurry base64 image. */}
+          {image.base64 && (
+            <Img src={image.base64} {...placeholderImageProps} />
+          )}
 
-            {/* Show the traced SVG image. */}
-            {image.tracedSVG && (
-              <Img
-                alt={alt}
-                title={title}
-                src={image.tracedSVG}
-                style={imagePlaceholderStyle}
-              />
-            )}
+          {/* Show the traced SVG image. */}
+          {image.tracedSVG && (
+            <Img src={image.tracedSVG} {...placeholderImageProps} />
+          )}
 
-            {/* Show a solid background color. */}
-            {bgColor && (
-              <Tag
-                title={title}
-                style={{
-                  backgroundColor: bgColor,
-                  position: `absolute`,
-                  top: 0,
-                  bottom: 0,
-                  opacity: !this.state.imgLoaded ? 1 : 0,
-                  transitionDelay: `0.35s`,
-                  right: 0,
-                  left: 0,
-                }}
-              />
-            )}
+          {/* Show a solid background color. */}
+          {bgColor && (
+            <Tag
+              title={title}
+              style={{
+                backgroundColor: bgColor,
+                position: `absolute`,
+                top: 0,
+                bottom: 0,
+                opacity: !this.state.imgLoaded ? 1 : 0,
+                transitionDelay: `0.35s`,
+                right: 0,
+                left: 0,
+              }}
+            />
+          )}
 
-            {/* Once the image is visible (or the browser doesn't support IntersectionObserver), start downloading the image */}
-            {this.state.isVisible && (
-              <picture>
-                {image.srcSetWebp && (<source
+          {/* Once the image is visible (or the browser doesn't support IntersectionObserver), start downloading the image */}
+          {this.state.isVisible && (
+            <picture>
+              {image.srcSetWebp && (
+                <source
                   type={`image/webp`}
                   srcSet={image.srcSetWebp}
                   sizes={image.sizes}
-                />)}
-
-                <source
-                  srcSet={image.srcSet}
-                  sizes={image.sizes}
                 />
+              )}
 
-                <Img
-                  alt={alt}
-                  title={title}
-                  src={image.src}
-                  style={imageStyle}
-                  ref={this.imageRef}
-                  onLoad={this.handleImageLoaded}
-                  onError={this.props.onError}
-                />
-              </picture>
-            )}
+              <source srcSet={image.srcSet} sizes={image.sizes} />
 
-            {/* Show the original image during server-side rendering if JavaScript is disabled */}
-            {this.state.hasNoScript && (
-              <noscript
-                dangerouslySetInnerHTML={{
-                  __html: noscriptImg({ alt, title, ...image }),
-                }}
+              <Img
+                alt={alt}
+                title={title}
+                src={image.src}
+                style={imageStyle}
+                ref={this.imageRef}
+                onLoad={this.handleImageLoaded}
+                onError={this.props.onError}
               />
-            )}
+            </picture>
+          )}
+
+          {/* Show the original image during server-side rendering if JavaScript is disabled */}
+          {this.state.hasNoScript && (
+            <noscript
+              dangerouslySetInnerHTML={{
+                __html: noscriptImg({ alt, title, ...image }),
+              }}
+            />
+          )}
         </Tag>
       )
     }
@@ -358,25 +358,16 @@ class Image extends React.Component {
           className={`${className ? className : ``} gatsby-image-wrapper`}
           style={divStyle}
           ref={this.handleRef}
+          key={`fixed-${JSON.stringify(image.srcSet)}`}
         >
           {/* Show the blurry base64 image. */}
           {image.base64 && (
-            <Img
-              alt={alt}
-              title={title}
-              src={image.base64}
-              style={imagePlaceholderStyle}
-            />
+            <Img src={image.base64} {...placeholderImageProps} />
           )}
 
           {/* Show the traced SVG image. */}
           {image.tracedSVG && (
-            <Img
-              alt={alt}
-              title={title}
-              src={image.tracedSVG}
-              style={imagePlaceholderStyle}
-            />
+            <Img src={image.tracedSVG} {...placeholderImageProps} />
           )}
 
           {/* Show a solid background color. */}
@@ -396,16 +387,15 @@ class Image extends React.Component {
           {/* Once the image is visible, start downloading the image */}
           {this.state.isVisible && (
             <picture>
-              {image.srcSetWebp && (<source
-                type={`image/webp`}
-                srcSet={image.srcSetWebp}
-                sizes={image.sizes}
-              />)}
+              {image.srcSetWebp && (
+                <source
+                  type={`image/webp`}
+                  srcSet={image.srcSetWebp}
+                  sizes={image.sizes}
+                />
+              )}
 
-              <source
-                srcSet={image.srcSet}
-                sizes={image.sizes}
-              />
+              <source srcSet={image.srcSet} sizes={image.sizes} />
 
               <Img
                 alt={alt}
@@ -485,6 +475,7 @@ Image.propTypes = {
   style: PropTypes.object,
   imgStyle: PropTypes.object,
   placeholderStyle: PropTypes.object,
+  placeholderClassName: PropTypes.string,
   backgroundColor: PropTypes.oneOfType([PropTypes.string, PropTypes.bool]),
   onLoad: PropTypes.func,
   onError: PropTypes.func,