+++ packages/gatsby-plugin-react-css-modules/src/gatsby-node.js
@@ -1,6 +1,8 @@
 exports.onCreateBabelConfig = ({ actions }, pluginOptions) => {
   const isDevelopment = process.env.NODE_ENV !== `production`
 
+  pluginOptions.plugins && delete pluginOptions.plugins
+
   actions.setBabelPlugin({
     name: `babel-plugin-react-css-modules`,
     options: {