+++ packages/gatsby/cache-dir/navigation.js
@@ -1,16 +1,11 @@
 import React from "react"
 import PropTypes from "prop-types"
-import loader, { setApiRunnerForLoader } from "./loader"
+import loader from "./loader"
 import redirects from "./redirects.json"
 import { apiRunner } from "./api-runner-browser"
 import emitter from "./emitter"
-import {
-  resolveRouteChangePromise,
-  resetRouteChangePromise,
-} from "./wait-for-route-change"
 import { navigate as reachNavigate } from "@reach/router"
 import parsePath from "./parse-path"
-import loadDirectlyOr404 from "./load-directly-or-404"
 
 // Convert to a map for faster lookup in maybeRedirect()
 const redirectMap = redirects.reduce((map, redirect) => {
@@ -48,7 +43,6 @@ const onPreRouteUpdate = location => {
 const onRouteUpdate = location => {
   if (!maybeRedirect(location.pathname)) {
     apiRunner(`onRouteUpdate`, { location })
-    resolveRouteChangePromise()
 
     // Temp hack while awaiting https://github.com/reach/router/issues/119
     window.__navigatingToLink = false
@@ -71,13 +65,15 @@ const navigate = (to, options = {}) => {
     pathname = parsePath(to).pathname
   }
 
-  // If we had a service worker update, no matter the path, reload window
+  // If we had a service worker update, no matter the path, reload window and
+  // reset the pathname whitelist
   if (window.GATSBY_SW_UPDATED) {
+    const { controller } = navigator.serviceWorker
+    controller.postMessage({ gatsbyApi: `resetWhitelist` })
+
     window.location = pathname
     return
   }
-
-  resetRouteChangePromise()
 
   // Start a timer to wait for a second before transitioning and showing a
   // loader in case resources aren't around yet.
@@ -89,26 +85,10 @@ const navigate = (to, options = {}) => {
   }, 1000)
 
   loader.getResourcesForPathname(pathname).then(pageResources => {
-    if (
-      (!pageResources || pageResources.page.path === `/404.html`) &&
-      process.env.NODE_ENV === `production`
-    ) {
-      clearTimeout(timeoutId)
-      loadDirectlyOr404(pageResources, to).then(() =>
-        reachNavigate(to, options)
-      )
-    } else {
-      reachNavigate(to, options)
-      clearTimeout(timeoutId)
-    }
+    reachNavigate(to, options)
+    clearTimeout(timeoutId)
   })
 }
-
-// reset route change promise after going back / forward
-// in history (when not using Gatsby navigation)
-window.addEventListener(`popstate`, () => {
-  resetRouteChangePromise()
-})
 
 function shouldUpdateScroll(prevRouterProps, { location }) {
   const { pathname, hash } = location
@@ -140,7 +120,6 @@ function init() {
   // Temp hack while awaiting https://github.com/reach/router/issues/119
   window.__navigatingToLink = false
 
-  setApiRunnerForLoader(apiRunner)
   window.___loader = loader
   window.___push = to => navigate(to, { replace: false })
   window.___replace = to => navigate(to, { replace: true })