+++ packages/gatsby-transformer-remark/CHANGELOG.md
@@ -3,6 +3,82 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.1.15"></a>
+
+## [2.1.15](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-remark/compare/gatsby-transformer-remark@2.1.14...gatsby-transformer-remark@2.1.15) (2018-11-30)
+
+### Bug Fixes
+
+- **gatsby-transformer-remark:** render html attributes in html excerpt ([#10196](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-remark/issues/10196)) ([2483aef](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-remark/commit/2483aef))
+
+<a name="2.1.14"></a>
+
+## [2.1.14](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-remark/compare/gatsby-transformer-remark@2.1.13...gatsby-transformer-remark@2.1.14) (2018-11-29)
+
+### Bug Fixes
+
+- **gatsby-transformer-remark:** properly bubble up errors thrown in subplugins ([#9972](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-remark/issues/9972)) ([b7d4656](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-remark/commit/b7d4656))
+
+<a name="2.1.13"></a>
+
+## [2.1.13](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-remark/compare/gatsby-transformer-remark@2.1.12...gatsby-transformer-remark@2.1.13) (2018-11-27)
+
+**Note:** Version bump only for package gatsby-transformer-remark
+
+<a name="2.1.12"></a>
+
+## [2.1.12](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-remark/compare/gatsby-transformer-remark@2.1.11...gatsby-transformer-remark@2.1.12) (2018-11-08)
+
+**Note:** Version bump only for package gatsby-transformer-remark
+
+<a name="2.1.11"></a>
+
+## [2.1.11](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-remark/compare/gatsby-transformer-remark@2.1.10...gatsby-transformer-remark@2.1.11) (2018-10-29)
+
+### Features
+
+- **gatsby:** Add nodes db module ([#9416](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-remark/issues/9416)) ([7d31fe7](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-remark/commit/7d31fe7)), closes [#9338](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-remark/issues/9338)
+
+<a name="2.1.10"></a>
+
+## [2.1.10](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-remark/compare/gatsby-transformer-remark@2.1.9...gatsby-transformer-remark@2.1.10) (2018-10-29)
+
+**Note:** Version bump only for package gatsby-transformer-remark
+
+<a name="2.1.9"></a>
+
+## [2.1.9](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-remark/compare/gatsby-transformer-remark@2.1.8...gatsby-transformer-remark@2.1.9) (2018-10-24)
+
+**Note:** Version bump only for package gatsby-transformer-remark
+
+<a name="2.1.8"></a>
+
+## [2.1.8](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-remark/compare/gatsby-transformer-remark@2.1.7...gatsby-transformer-remark@2.1.8) (2018-10-18)
+
+**Note:** Version bump only for package gatsby-transformer-remark
+
+<a name="2.1.7"></a>
+
+## [2.1.7](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-remark/compare/gatsby-transformer-remark@2.1.6...gatsby-transformer-remark@2.1.7) (2018-10-09)
+
+### Features
+
+- add error message with filename on Markdown error, fix bug in panicOnBuild ([#8866](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-remark/issues/8866)) ([bbff3be](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-remark/commit/bbff3be))
+
+<a name="2.1.6"></a>
+
+## [2.1.6](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-remark/compare/gatsby-transformer-remark@2.1.5...gatsby-transformer-remark@2.1.6) (2018-10-02)
+
+### Bug Fixes
+
+- support path-prefix for reference links in Markdown files ([#8607](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-remark/issues/8607)) ([fb43fda](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-remark/commit/fb43fda)), closes [#8588](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-remark/issues/8588)
+
+<a name="2.1.5"></a>
+
+## [2.1.5](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-remark/compare/gatsby-transformer-remark@2.1.4...gatsby-transformer-remark@2.1.5) (2018-10-01)
+
+**Note:** Version bump only for package gatsby-transformer-remark
+
 <a name="2.1.4"></a>
 
 ## [2.1.4](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-remark/compare/gatsby-transformer-remark@2.1.3...gatsby-transformer-remark@2.1.4) (2018-09-26)